Manually separate mockgens
